refactor(cmd): extract command/flag wiring into endpoint_flags.go
Move the endpoint init() (command-tree assembly via AddCommand plus all
Flags().XxxVar registration) and the registerSplunkFlags/registerFalconFlags
helpers out of endpoint.go into a new same-package endpoint_flags.go.

endpoint.go now holds the command and option-struct declarations plus the small
run/helper functions (~577 lines, down from 768). Move-only, behavior-preserving:
cmd tests pass.
