+++ cli/beacon/cmd/endpoint.go
@@ -10,7 +10,6 @@ import (
 	"github.com/spf13/cobra"
 
 	endpointconfig "github.com/asymptote-labs/agent-beacon/cli/beacon/internal/endpoint/config"
-	"github.com/asymptote-labs/agent-beacon/cli/beacon/internal/endpoint/dashboard"
 	"github.com/asymptote-labs/agent-beacon/cli/beacon/internal/endpoint/integrations/cowork"
 	"github.com/asymptote-labs/agent-beacon/cli/beacon/internal/endpoint/integrations/openclaw"
 	"github.com/asymptote-labs/agent-beacon/cli/beacon/internal/endpoint/integrations/vscode"
@@ -373,176 +372,6 @@ var endpointVSCodeValidateCmd = &cobra.Command{
 	RunE:         func(cmd *cobra.Command, args []string) error { return runEndpointVSCodeValidate() },
 }
 
-func init() {
-	rootCmd.AddCommand(endpointCmd)
-	rootCmd.AddCommand(topLevelDoctorCmd)
-	rootCmd.AddCommand(topLevelStatusCmd)
-	rootCmd.AddCommand(topLevelInventoryCmd)
-
-	endpointCmd.AddCommand(endpointInstallCmd)
-	endpointCmd.AddCommand(endpointStatusCmd)
-	endpointCmd.AddCommand(endpointDoctorCmd)
-	endpointCmd.AddCommand(endpointInventoryCmd)
-	endpointCmd.AddCommand(endpointDiscoverCmd)
-	endpointCmd.AddCommand(endpointTestEventCmd)
-	endpointCmd.AddCommand(endpointBundleDiagnosticsCmd)
-	endpointCmd.AddCommand(endpointUninstallCmd)
-	endpointCmd.AddCommand(endpointRepairCmd)
-	endpointCmd.AddCommand(endpointDashboardCmd)
-	for _, c := range buildDestinationCommands() {
-		endpointCmd.AddCommand(c)
-	}
-	endpointCmd.AddCommand(endpointIntegrationsCmd)
-	endpointCmd.AddCommand(endpointHooksCmd)
-	endpointCmd.AddCommand(endpointConfigCmd)
-	endpointInventoryCmd.AddCommand(endpointInventoryHeartbeatCmd)
-	endpointConfigCmd.AddCommand(endpointConfigShowCmd)
-	endpointConfigCmd.AddCommand(endpointConfigValidateCmd)
-	endpointIntegrationsCmd.AddCommand(endpointIntegrationsValidateCmd)
-	endpointIntegrationsCmd.AddCommand(endpointCoworkCmd)
-	endpointIntegrationsCmd.AddCommand(endpointOpenClawCmd)
-	endpointIntegrationsCmd.AddCommand(endpointVSCodeCmd)
-	endpointHooksCmd.AddCommand(endpointHooksInstallCmd)
-	endpointHooksCmd.AddCommand(endpointHooksUninstallCmd)
-	endpointHooksCmd.AddCommand(endpointHooksStatusCmd)
-	endpointCoworkCmd.AddCommand(endpointCoworkPrintConfigCmd)
-	endpointCoworkCmd.AddCommand(endpointCoworkSetupCmd)
-	endpointCoworkCmd.AddCommand(endpointCoworkStatusCmd)
-	endpointCoworkCmd.AddCommand(endpointCoworkValidateCmd)
-	endpointOpenClawCmd.AddCommand(endpointOpenClawPrintConfigCmd)
-	endpointOpenClawCmd.AddCommand(endpointOpenClawStatusCmd)
-	endpointOpenClawCmd.AddCommand(endpointOpenClawValidateCmd)
-	endpointVSCodeCmd.AddCommand(endpointVSCodePrintConfigCmd)
-	endpointVSCodeCmd.AddCommand(endpointVSCodeSetupCmd)
-	endpointVSCodeCmd.AddCommand(endpointVSCodeStatusCmd)
-	endpointVSCodeCmd.AddCommand(endpointVSCodeValidateCmd)
-
-	for _, c := range []*cobra.Command{endpointInstallCmd, endpointStatusCmd, endpointDoctorCmd, endpointInventoryCmd, endpointInventoryHeartbeatCmd, endpointDiscoverCmd, endpointTestEventCmd, endpointBundleDiagnosticsCmd, endpointUninstallCmd, endpointRepairCmd, endpointConfigShowCmd, endpointConfigValidateCmd, endpointIntegrationsValidateCmd, topLevelDoctorCmd, topLevelStatusCmd, topLevelInventoryCmd} {
-		c.Flags().BoolVar(&endpointOpts.userMode, "user", true, "Use per-user endpoint paths")
-		c.Flags().BoolVar(&endpointOpts.systemMode, "system", false, "Use system endpoint paths and launch daemon")
-		c.Flags().StringVar(&endpointOpts.logPath, "log-path", "", "Runtime JSONL log path")
-	}
-
-	endpointInstallCmd.Flags().StringVar(&endpointOpts.harnesses, "harness", "claude,codex", "Comma-separated harnesses to configure")
-	endpointInstallCmd.Flags().IntVar(&endpointOpts.grpcPort, "otlp-grpc-port", endpointconfig.DefaultGRPCPort, "Local OTLP gRPC port")
-	endpointInstallCmd.Flags().IntVar(&endpointOpts.httpPort, "otlp-http-port", endpointconfig.DefaultHTTPPort, "Local OTLP HTTP port")
-	endpointInstallCmd.Flags().StringVar(&endpointOpts.collectorPath, "collector", "", "Path to a beacon-otelcol binary")
-	endpointInstallCmd.Flags().BoolVar(&endpointOpts.includeRuntimeMetrics, "include-runtime-metrics", false, "Include generic process/runtime OTLP metrics and harness operational metrics (OpenClaw, Copilot CLI) in the runtime JSONL log")
-	endpointInstallCmd.Flags().BoolVar(&endpointOpts.includeCodexSpans, "include-codex-spans", false, "Include high-volume Codex OTLP spans for troubleshooting")
-	endpointInstallCmd.Flags().BoolVar(&endpointOpts.noStart, "no-start", false, "Write files without starting the launchd service")
-	endpointInstallCmd.Flags().BoolVar(&endpointOpts.dryRun, "dry-run", false, "Print planned actions without changing endpoint files or services")
-	endpointInstallCmd.Flags().StringVar(&endpointOpts.contentRetention, "content-retention", "", "Deprecated no-op; Beacon always captures full content subject to redaction and size limits")
-	_ = endpointInstallCmd.Flags().MarkHidden("content-retention")
-	_ = endpointInstallCmd.Flags().MarkDeprecated("content-retention", "Beacon now always captures full content; this flag is ignored")
-	registerSplunkFlags(endpointInstallCmd)
-	registerFalconFlags(endpointInstallCmd)
-	endpointRepairCmd.Flags().StringVar(&endpointOpts.harnesses, "harness", "claude,codex", "Comma-separated harnesses to configure")
-	endpointRepairCmd.Flags().IntVar(&endpointOpts.grpcPort, "otlp-grpc-port", endpointconfig.DefaultGRPCPort, "Local OTLP gRPC port")
-	endpointRepairCmd.Flags().IntVar(&endpointOpts.httpPort, "otlp-http-port", endpointconfig.DefaultHTTPPort, "Local OTLP HTTP port")
-	endpointRepairCmd.Flags().StringVar(&endpointOpts.collectorPath, "collector", "", "Path to a beacon-otelcol binary")
-	endpointRepairCmd.Flags().BoolVar(&endpointOpts.includeRuntimeMetrics, "include-runtime-metrics", false, "Include generic process/runtime OTLP metrics and harness operational metrics (OpenClaw, Copilot CLI) in the runtime JSONL log")
-	endpointRepairCmd.Flags().BoolVar(&endpointOpts.includeCodexSpans, "include-codex-spans", false, "Include high-volume Codex OTLP spans for troubleshooting")
-	endpointRepairCmd.Flags().BoolVar(&endpointOpts.noStart, "no-start", false, "Write files without starting the launchd service")
-	endpointRepairCmd.Flags().BoolVar(&endpointOpts.dryRun, "dry-run", false, "Print planned actions without changing endpoint files or services")
-	endpointRepairCmd.Flags().StringVar(&endpointOpts.contentRetention, "content-retention", "", "Deprecated no-op; Beacon always captures full content subject to redaction and size limits")
-	_ = endpointRepairCmd.Flags().MarkHidden("content-retention")
-	_ = endpointRepairCmd.Flags().MarkDeprecated("content-retention", "Beacon now always captures full content; this flag is ignored")
-	registerSplunkFlags(endpointRepairCmd)
-	registerFalconFlags(endpointRepairCmd)
-	endpointDashboardCmd.Flags().BoolVar(&endpointOpts.userMode, "user", true, "Use per-user endpoint paths")
-	endpointDashboardCmd.Flags().BoolVar(&endpointOpts.systemMode, "system", false, "Use system endpoint paths and launch daemon")
-	endpointDashboardCmd.Flags().StringVar(&endpointOpts.logPath, "log-path", "", "Runtime JSONL log path")
-	endpointDashboardCmd.Flags().StringVar(&dashboardOpts.addr, "addr", dashboard.DefaultAddr, "Local dashboard listen address")
-	endpointDashboardCmd.Flags().BoolVar(&dashboardOpts.open, "open", false, "Open the dashboard in a browser")
-
-	endpointDiscoverCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print discovery as JSON")
-	endpointDiscoverCmd.Flags().BoolVar(&endpointOpts.allTargets, "all", false, "Discover all supported runtime targets")
-	endpointStatusCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print status as JSON")
-	endpointDoctorCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print doctor results as JSON")
-	endpointDoctorCmd.Flags().BoolVar(&endpointOpts.fix, "fix", false, "Apply safe endpoint doctor remediations")
-	endpointInventoryCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print inventory as JSON")
-	endpointInventoryCmd.Flags().BoolVar(&endpointOpts.allTargets, "all", false, "Include all supported targets")
-	endpointInventoryCmd.Flags().BoolVar(&inventoryOpts.mcp, "mcp", false, "Show only MCP server inventory and source configs")
-	endpointInventoryCmd.Flags().BoolVar(&inventoryOpts.skills, "skills", false, "Show only local agent skill inventory")
-	endpointInventoryCmd.Flags().BoolVar(&inventoryOpts.hooks, "hooks", false, "Show only hook integration and hook config inventory")
-	endpointInventoryCmd.Flags().BoolVar(&inventoryOpts.writeEvent, "write-event", false, "Append inventory events to the endpoint runtime log")
-	endpointInventoryHeartbeatCmd.Flags().BoolVar(&inventoryOpts.heartbeatForce, "force", false, "Write inventory heartbeat even when TTL has not expired")
-	endpointInventoryHeartbeatCmd.Flags().StringVar(&inventoryOpts.heartbeatConfig, "config", "", "Endpoint config path")
-	endpointInventoryHeartbeatCmd.Flags().StringVar(&inventoryOpts.workingDir, "working-dir", "", "Working directory for project inventory")
-	endpointInventoryHeartbeatCmd.Flags().StringVar(&inventoryOpts.trigger, "trigger", "manual", "Inventory trigger source")
-	endpointInventoryHeartbeatCmd.Flags().StringVar(&inventoryOpts.triggerHarness, "trigger-harness", "", "Harness that triggered inventory")
-	endpointInventoryHeartbeatCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print heartbeat result as JSON")
-	topLevelDoctorCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print doctor results as JSON")
-	topLevelDoctorCmd.Flags().BoolVar(&endpointOpts.fix, "fix", false, "Apply safe endpoint doctor remediations")
-	topLevelStatusCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print status as JSON")
-	topLevelInventoryCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print inventory as JSON")
-	topLevelInventoryCmd.Flags().BoolVar(&endpointOpts.allTargets, "all", false, "Include all supported targets")
-	topLevelInventoryCmd.Flags().BoolVar(&inventoryOpts.mcp, "mcp", false, "Show only MCP server inventory and source configs")
-	topLevelInventoryCmd.Flags().BoolVar(&inventoryOpts.skills, "skills", false, "Show only local agent skill inventory")
-	topLevelInventoryCmd.Flags().BoolVar(&inventoryOpts.hooks, "hooks", false, "Show only hook integration and hook config inventory")
-	topLevelInventoryCmd.Flags().BoolVar(&inventoryOpts.writeEvent, "write-event", false, "Append inventory events to the endpoint runtime log")
-	endpointTestEventCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print validation stages as JSON")
-	endpointBundleDiagnosticsCmd.Flags().StringVar(&endpointOpts.outputDir, "output", "", "Output directory for diagnostics bundle")
-	endpointBundleDiagnosticsCmd.Flags().BoolVar(&endpointOpts.includeEventSummaries, "include-event-summaries", false, "Include redacted event summaries")
-	endpointBundleDiagnosticsCmd.Flags().BoolVar(&endpointOpts.includeRawEvents, "include-raw-events", false, "Include raw runtime JSONL events")
-	endpointIntegrationsValidateCmd.Flags().BoolVar(&endpointOpts.allTargets, "all", false, "Validate all supported admin integrations")
-	endpointIntegrationsValidateCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print validation as JSON")
-	for _, c := range []*cobra.Command{endpointCoworkPrintConfigCmd, endpointCoworkSetupCmd, endpointCoworkStatusCmd, endpointCoworkValidateCmd} {
-		c.Flags().BoolVar(&endpointOpts.userMode, "user", true, "Use per-user endpoint paths")
-		c.Flags().BoolVar(&endpointOpts.systemMode, "system", false, "Use system endpoint paths and launch daemon")
-		c.Flags().StringVar(&endpointOpts.logPath, "log-path", "", "Runtime JSONL log path")
-	}
-	endpointCoworkPrintConfigCmd.Flags().StringVar(&coworkOpts.headers, "headers", "", "Optional OTLP headers to show in setup guidance")
-	endpointCoworkPrintConfigCmd.Flags().StringVar(&coworkOpts.endpoint, "endpoint", "", "Public OTLP HTTPS endpoint to show in setup guidance")
-	endpointCoworkPrintConfigCmd.Flags().StringVar(&coworkOpts.resourceAttributes, "resource-attributes", "", "Optional Claude Cowork resource attributes")
-	endpointCoworkSetupCmd.Flags().StringVar(&coworkOpts.endpoint, "endpoint", "", "Public OTLP HTTPS endpoint reachable by Claude Cowork")
-	endpointCoworkSetupCmd.Flags().StringVar(&coworkOpts.headers, "headers", "", "Optional OTLP headers for the Claude admin settings")
-	endpointCoworkSetupCmd.Flags().StringVar(&coworkOpts.resourceAttributes, "resource-attributes", "", "Optional Claude Cowork resource attributes")
-	endpointCoworkSetupCmd.Flags().BoolVar(&coworkOpts.ngrok, "ngrok", false, "Create a temporary authenticated ngrok tunnel to the local OTLP HTTP receiver")
-	endpointCoworkSetupCmd.Flags().BoolVar(&coworkOpts.open, "open", false, "Open Claude Cowork admin settings in a browser")
-	endpointCoworkValidateCmd.Flags().StringVar(&coworkOpts.headers, "headers", "", "Optional OTLP headers to show when validation fails")
-	endpointCoworkValidateCmd.Flags().StringVar(&coworkOpts.endpoint, "endpoint", "", "Public OTLP HTTPS endpoint to show when validation fails")
-	endpointCoworkValidateCmd.Flags().StringVar(&coworkOpts.resourceAttributes, "resource-attributes", "", "Optional Claude Cowork resource attributes")
-	endpointCoworkValidateCmd.Flags().StringVar(&coworkOpts.since, "since", "", "Require a Claude Cowork event within this duration, such as 10m")
-	endpointCoworkStatusCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print status as JSON")
-	for _, c := range []*cobra.Command{endpointOpenClawPrintConfigCmd, endpointOpenClawStatusCmd, endpointOpenClawValidateCmd} {
-		c.Flags().BoolVar(&endpointOpts.userMode, "user", true, "Use per-user endpoint paths")
-		c.Flags().BoolVar(&endpointOpts.systemMode, "system", false, "Use system endpoint paths and launch daemon")
-		c.Flags().StringVar(&endpointOpts.logPath, "log-path", "", "Runtime JSONL log path")
-	}
-	endpointOpenClawPrintConfigCmd.Flags().StringVar(&openClawOpts.endpoint, "endpoint", "", "OTLP HTTP endpoint to show in setup guidance")
-	endpointOpenClawValidateCmd.Flags().StringVar(&openClawOpts.endpoint, "endpoint", "", "OTLP HTTP endpoint to show when validation fails")
-	endpointOpenClawValidateCmd.Flags().StringVar(&openClawOpts.since, "since", "", "Require an OpenClaw event within this duration, such as 10m")
-	endpointOpenClawStatusCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print status as JSON")
-	for _, c := range []*cobra.Command{endpointVSCodePrintConfigCmd, endpointVSCodeSetupCmd, endpointVSCodeStatusCmd, endpointVSCodeValidateCmd} {
-		c.Flags().BoolVar(&endpointOpts.userMode, "user", true, "Use per-user endpoint paths")
-		c.Flags().BoolVar(&endpointOpts.systemMode, "system", false, "Use system endpoint paths and launch daemon")
-		c.Flags().StringVar(&endpointOpts.logPath, "log-path", "", "Runtime JSONL log path")
-		c.Flags().StringVar(&vscodeOpts.endpoint, "endpoint", "", "OTLP HTTP endpoint for VS Code Copilot")
-		c.Flags().StringVar(&vscodeOpts.workspace, "workspace", "", "Workspace path for .vscode/settings.json")
-		c.Flags().BoolVar(&vscodeOpts.captureContent, "capture-content", false, "Enable full Copilot prompt, response, tool argument, and tool result capture")
-	}
-	endpointVSCodeSetupCmd.Flags().BoolVar(&endpointOpts.dryRun, "dry-run", false, "Print VS Code setup guidance without changing settings")
-	endpointVSCodeValidateCmd.Flags().StringVar(&vscodeOpts.since, "since", "", "Require a VS Code event within this duration, such as 10m")
-	endpointVSCodeStatusCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print status as JSON")
-	for _, c := range []*cobra.Command{endpointHooksInstallCmd, endpointHooksUninstallCmd, endpointHooksStatusCmd} {
-		c.Flags().BoolVar(&endpointOpts.userMode, "user", true, "Use per-user endpoint paths")
-		c.Flags().BoolVar(&endpointOpts.systemMode, "system", false, "Use system endpoint paths and launch daemon")
-		c.Flags().StringVar(&endpointOpts.logPath, "log-path", "", "Runtime JSONL log path")
-		c.Flags().StringVar(&endpointOpts.hookHarnesses, "harness", "cursor", "Comma-separated hook harnesses")
-		c.Flags().StringVar(&endpointOpts.hookLevel, "level", "user", "Hook install level: user or project")
-	}
-	endpointHooksInstallCmd.Flags().BoolVar(&endpointOpts.allTargets, "all", false, "Target all supported hook harnesses")
-	endpointHooksInstallCmd.Flags().BoolVar(&endpointOpts.dryRun, "dry-run", false, "Print planned hook actions without changing files")
-	endpointHooksUninstallCmd.Flags().BoolVar(&endpointOpts.allTargets, "all", false, "Target all supported hook harnesses")
-	endpointHooksUninstallCmd.Flags().BoolVar(&endpointOpts.dryRun, "dry-run", false, "Print planned hook actions without changing files")
-	endpointHooksStatusCmd.Flags().BoolVar(&endpointOpts.jsonOutput, "json", false, "Print status as JSON")
-	endpointHooksStatusCmd.Flags().BoolVar(&endpointOpts.allTargets, "all", false, "Show all supported hook harnesses")
-	endpointUninstallCmd.Flags().BoolVar(&endpointOpts.keepLogs, "keep-logs", false, "Keep runtime logs during uninstall")
-	endpointUninstallCmd.Flags().BoolVar(&endpointOpts.keepConfig, "keep-config", false, "Keep harness telemetry configuration during uninstall")
-	endpointUninstallCmd.Flags().BoolVar(&endpointOpts.dryRun, "dry-run", false, "Print planned actions without changing endpoint files or services")
-}
-
 func runEndpointOpenClawValidate() error {
 	cfg := loadOrDefaultConfig()
 	setup := func() {
@@ -703,26 +532,6 @@ func splitHarnessCSV(value string) []string {
 		return []string{}
 	}
 	return splitCSV(value)
-}
-
-func registerSplunkFlags(cmd *cobra.Command) {
-	cmd.Flags().StringVar(&splunkOpts.hecEndpoint, "splunk-hec-endpoint", "", "Splunk HEC endpoint URL")
-	cmd.Flags().StringVar(&splunkOpts.hecToken, "splunk-hec-token", "", "Splunk HEC token")
-	cmd.Flags().StringVar(&splunkOpts.index, "splunk-index", "", "Optional Splunk index")
-	cmd.Flags().StringVar(&splunkOpts.source, "splunk-source", endpointconfig.DefaultSplunkSource, "Optional Splunk source")
-	cmd.Flags().StringVar(&splunkOpts.sourcetype, "splunk-sourcetype", endpointconfig.DefaultSplunkSourcetype, "Optional Splunk sourcetype")
-	cmd.Flags().BoolVar(&splunkOpts.insecureSkipVerify, "splunk-insecure-skip-verify", false, "Skip Splunk HEC TLS certificate verification")
-	cmd.Flags().StringVar(&splunkOpts.caFile, "splunk-ca-file", "", "Optional CA certificate path for Splunk HEC TLS verification")
-}
-
-func registerFalconFlags(cmd *cobra.Command) {
-	cmd.Flags().StringVar(&falconOpts.hecEndpoint, "falcon-hec-endpoint", "", "Falcon LogScale HEC endpoint URL")
-	cmd.Flags().StringVar(&falconOpts.hecToken, "falcon-hec-token", "", "Falcon LogScale ingest token")
-	cmd.Flags().StringVar(&falconOpts.index, "falcon-index", "", "Optional Falcon LogScale repository for multi-repository tokens")
-	cmd.Flags().StringVar(&falconOpts.source, "falcon-source", endpointconfig.DefaultFalconSource, "Optional Falcon LogScale source")
-	cmd.Flags().StringVar(&falconOpts.sourcetype, "falcon-sourcetype", endpointconfig.DefaultFalconSourcetype, "Optional Falcon LogScale parser or sourcetype")
-	cmd.Flags().BoolVar(&falconOpts.insecureSkipVerify, "falcon-insecure-skip-verify", false, "Skip Falcon LogScale HEC TLS certificate verification")
-	cmd.Flags().StringVar(&falconOpts.caFile, "falcon-ca-file", "", "Optional CA certificate path for Falcon LogScale HEC TLS verification")
 }
 
 func splunkHECOptions() *endpointconfig.SplunkHEC {
